:hammer: [chore] Component 폴더 위치 수정 #70
