+++ Yeowoo/Sources/Component/CardViewModifier.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 
 struct CardViewModifier: ViewModifier {
-    //default 값 66 설정
+
     var height: CGFloat = 66
     
     func body(content: Content) -> some View {
@@ -18,8 +18,5 @@ struct CardViewModifier: ViewModifier {
             .background(RoundedRectangle(cornerRadius: 10, style: .continuous)
                 .fill(.white))
             .shadow(color: Color.black.opacity(0.05), radius: 20, x: 0, y: 4)
-            
-
     }
 }
-
