+++ Yeowoo/Sources/Views/Album/InviteFriendView.swift
@@ -35,7 +35,7 @@ struct InviteFriendView: View {
 				.padding(. leading, 30)
 			
 			ZStack {
-				GrayTitleMakingView(placeholder: "친구 아이디를 입력해주세요", text: $friendID)
+				GrayTitleMakingView(text: $friendID, placeholder: "친구 아이디를 입력해주세요")
 				
 				HStack {
 					Spacer()
@@ -61,7 +61,7 @@ struct InviteFriendView: View {
 			
 			ScrollView{
 				ForEach(myFriend.indices, id: \.self) { index in
-					FindFriendContents(user: myFriend[index], friendToggle: $friendToggles[index])
+					FindFriendContents(friendToggle: $friendToggles[index], user: myFriend[index])
 				}
 				
 				Spacer()
